# apps/marketplace/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace Simulator API
  version: 1.0.0
  description: |
    A provider-aware marketplace API. Order and fulfillment contracts are only
    available through `SHOPEE_LIKE` and `TOKOPEDIA_LIKE`; Generic is not an
    order-provider profile.

    Shared catalogue, warehouse, and webhook requests under `/api/v1` are HMAC-SHA256 signed using `METHOD + PATH + UNIX_TIMESTAMP + RAW_BODY`.
    Every mutating public operation across the shared, Shopee-like, and
    Tokopedia-like contracts require an `Idempotency-Key`. Concurrent use is
    claimed atomically, completed responses are replayed, and reusing a key for
    a different logical request returns `409`.

    Successful and rate-limited public responses expose `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, and `X-RateLimit-Reset` (Unix seconds). Public requests can return
    `503 MARKETPLACE_MAINTENANCE` when the administrator enables global maintenance mode.

    Webhooks are asynchronous. A 2xx response succeeds; all other responses and network failures
    retry at 30 seconds, 2 minutes, 10 minutes, and 30 minutes. Delivery attempts are immutable.
    Per-shop scenarios can inject slow responses, random 500s, timeouts, rate limits, duplicate,
    delayed, out-of-order, and failed webhook deliveries without affecting other shops.

    `SHOPEE_LIKE` shops additionally expose `/api/shopee/v1`. This is a distinct
...
    `X-Shopee-Signature`, where the signature is HMAC-SHA256 of
...

...
servers:
...
paths:
...
  /api/shopee/v1/orders:
    get:
      operationId: shopeeListOrders
      tags: [Orders]
      summary: List SHOPEE_LIKE orders with page-number pagination
      parameters: [{ $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }, { name: page_no, in: query, schema: { type: integer, minimum: 1, default: 1 } }, { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }, { name: order_status, in: query, schema: { type: string } }, { name: time_from, in: query, description: Unix seconds, schema: { type: integer } }, { name: time_to, in: query, description: Unix seconds, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '400': { $ref: '#/components/responses/ShopeeError' }, '401': { $ref: '#/components/responses/ShopeeError' }, '403': { $ref: '#/components/responses/ShopeeError' }, '429': { $ref: '#/components/responses/ShopeeError' } }
...
  /api/shopee/v1/orders/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get:
      operationId: shopeeGetOrder
      tags: [Orders]
      summary: Get a SHOPEE_LIKE order using provider terminology
      parameters: [{ $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }]
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '401': { $ref: '#/components/responses/ShopeeError' }, '404': { $ref: '#/components/responses/ShopeeError' } }
  /api/shopee/v1/orders/{id}/cancel:
    post:
      operationId: shopeeCancelOrder
      tags: [Orders]
      summary: Customer cancellation with Shopee-like request/error contract
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [cancel_reason], properties: { cancel_reason: { type: string, enum: [CHANGE_OF_MIND, DUPLICATE_ORDER, ADDRESS_ISSUE] } } } } } }
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '400': { $ref: '#/components/responses/ShopeeError' }, '401': { $ref: '#/components/responses/ShopeeError' }, '409': { $ref: '#/components/responses/ShopeeError' }, '413': { $ref: '#/components/responses/ShopeeError' } }
  /api/shopee/v1/orders/{id}/ship-order:
    post:
      operationId: shopeeProcessOrder
      tags: [Orders]
      summary: Move a paid SHOPEE_LIKE order into seller processing
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '400': { $ref: '#/components/responses/ShopeeError' }, '401': { $ref: '#/components/responses/ShopeeError' }, '409': { $ref: '#/components/responses/ShopeeError' }, '413': { $ref: '#/components/responses/ShopeeError' } }
...
  /api/shopee/v1/orders/{id}/packages:
    post:
      operationId: shopeeCreatePackage
      tags: [Fulfillment]
      summary: Allocate selected order items into a Shopee-like package
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PackageInput' } } } }
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '400': { $ref: '#/components/responses/ShopeeError' }, '401': { $ref: '#/components/responses/ShopeeError' }, '409': { $ref: '#/components/responses/ShopeeError' }, '413': { $ref: '#/components/responses/ShopeeError' } }
  /api/shopee/v1/orders/{id}/shipments:
    post:
      operationId: shopeeCreateShipment
      tags: [Fulfillment]
      summary: Create a Shopee-like shipment for a ready-to-ship order
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ShipmentInput' } } } }
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '400': { $ref: '#/components/responses/ShopeeError' }, '401': { $ref: '#/components/responses/ShopeeError' }, '409': { $ref: '#/components/responses/ShopeeError' }, '413': { $ref: '#/components/responses/ShopeeError' } }
...
  /api/tokopedia/v202309/orders/{id}/cancel:
    post:
      operationId: tokopediaCancelOrder
      tags: [Tokopedia Like]
      summary: Cancel an order with the Tokopedia-like contract
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/TokopediaAppKey' }, { $ref: '#/components/parameters/TokopediaTimestamp' }, { $ref: '#/components/parameters/TokopediaSign' }, { $ref: '#/components/parameters/TokopediaAccessToken' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string, enum: [CHANGE_OF_MIND, DUPLICATE_ORDER, ADDRESS_ISSUE] } } } } } }
      responses: { '200': { $ref: '#/components/responses/TokopediaResponse' }, '400': { $ref: '#/components/responses/TokopediaError' }, '401': { $ref: '#/components/responses/TokopediaError' }, '409': { $ref: '#/components/responses/TokopediaError' }, '413': { $ref: '#/components/responses/TokopediaError' }, '429': { $ref: '#/components/responses/TokopediaError' } }
  /api/tokopedia/v202309/orders/{id}/shipments:
    post:
      operationId: tokopediaCreateShipment
      tags: [Tokopedia Like, Fulfillment]
      summary: Create a shipment through the Tokopedia-like fulfillment contract
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/TokopediaAppKey' }, { $ref: '#/components/parameters/TokopediaTimestamp' }, { $ref: '#/components/parameters/TokopediaSign' }, { $ref: '#/components/parameters/TokopediaAccessToken' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ShipmentInput' } } } }
      responses: { '200': { $ref: '#/components/responses/TokopediaResponse' }, '400': { $ref: '#/components/responses/TokopediaError' }, '401': { $ref: '#/components/responses/TokopediaError' }, '409': { $ref: '#/components/responses/TokopediaError' }, '413': { $ref: '#/components/responses/TokopediaError' }, '429': { $ref: '#/components/responses/TokopediaError' } }
...
components:
  securitySchemes:
    ControlPlaneBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
...
    ShopeePartnerID: { name: X-Shopee-Partner-Id, in: header, required: true, schema: { type: string }, description: Credential client ID used as the provider partner ID. }
    ShopeeTimestamp: { name: X-Shopee-Timestamp, in: header, required: true, schema: { type: string, description: Unix seconds. } }
    ShopeeSignature: { name: X-Shopee-Signature, in: header, required: true, schema: { type: string, description: Hex HMAC-SHA256 of PARTNER_ID + PATH + TIMESTAMP + raw body. } }
    TokopediaAppKey: { name: app_key, in: query, required: true, schema: { type: string }, description: Partner Center app key (the control-plane credential client ID). }
    TokopediaTimestamp: { name: timestamp, in: query, required: true, schema: { type: integer, format: int64 }, description: Unix seconds; accepted within five minutes. }
    TokopediaSign: { name: sign, in: query, required: true, schema: { type: string }, description: Hex HMAC-SHA256 of APP_SECRET + PATH + sorted query excluding sign/access_token + raw body + APP_SECRET. }
    TokopediaAccessToken: { name: x-tts-access-token, in: header, required: true, schema: { type: string }, description: Access token returned only when the control-plane credential is created. }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, maxLength: 255 }, description: Stable key for one logical mutation. Retrying the identical method, path, stable query, and body replays its completed response; a different payload returns 409. }
...
  schemas:
...
    ShipmentInput:
      type: object
      required: [shipping_provider, pickup_type]
      properties:
        shipping_provider: { type: string, example: provider_express }
        pickup_type: { type: string, enum: [PICKUP] }
        package_id: { type: string, description: Existing package allocated for this order. }
    PackageInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [order_item_id, quantity]
            properties:
              order_item_id: { type: string }
              quantity: { type: integer, minimum: 1 }
...
  responses:
    ShopeeResponse: { description: Successful SHOPEE_LIKE response, headers: { X-Shopee-Api-Call-Limit: { schema: { type: string }, description: Calls used/current limit. }, X-Shopee-RateLimit-Reset: { schema: { type: integer }, description: Unix reset time. } }, content: { application/json: { schema: { type: object, required: [error, message, request_id, response], properties: { error: { type: string, example: '' }, message: { type: string, example: success }, request_id: { type: string }, response: { type: object } } } } } }
    ShopeeError: { description: SHOPEE_LIKE error, content: { application/json: { schema: { type: object, required: [error, message, request_id], properties: { error: { type: string, example: error_invalid_state }, message: { type: string }, request_id: { type: string } } } } } }
    TokopediaResponse: { description: Successful TOKOPEDIA_LIKE response, headers: { X-TTS-Api-Call-Limit: { schema: { type: integer } }, X-TTS-RateLimit-Reset: { schema: { type: integer, description: Unix reset time. } } }, content: { application/json: { schema: { type: object, required: [code, message, request_id, data], properties: { code: { type: integer, example: 0 }, message: { type: string, example: Success }, request_id: { type: string }, data: { type: object } } } } } }
    TokopediaError: { description: TOKOPEDIA_LIKE error, content: { application/json: { schema: { type: object, required: [code, message, request_id, data], properties: { code: { type: integer, example: 36000003 }, message: { type: string }, request_id: { type: string }, data: { type: object } } } } } }
...